requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>-3.9.0,<3.10.0
psycopg2>=2.7.5,<2.8.0
django-extensions>=2.2.9,<2.3.0
mapbox-vector-tile>=1.2.0,<1.3.0
djangorestframework-gis>=0.0.0,<0.2.0